requirements.txt
Flask
python-dotenv
watchdog
Flask-JSON
flask-cors
requests
beautifulsoup4